Fix map sometimes not being loaded properly
This was the bug that took ages to fix. It is due to the fact that the read() system call does not guarantee it will read in the specified number of bytes. Rather, it will read up to a maximum of the number of bytes specified. How many bytes it actually read depended on the system, which is why it worked on some computers but not others.